# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Creates a new project request. The server validates inputs, persists the
order, and (for PIX) generates the dynamic BR Code payload + QR image.
The amount is server-controlled via the validated input — never trust a
client-provided BR Code.

 * @summary Create a new project order
 */
export const createOrderBodyClientNameMin = 2;
export const createOrderBodyClientNameMax = 80;

export const createOrderBodyContactEmailMax = 254;

export const createOrderBodyTitleMin = 4;
export const createOrderBodyTitleMax = 120;

export const createOrderBodyDescriptionMin = 10;
export const createOrderBodyDescriptionMax = 2000;

export const CreateOrderBody = zod.object({
  clientName: zod
    .string()
    .min(createOrderBodyClientNameMin)
    .max(createOrderBodyClientNameMax),
  contactEmail: zod.string().email().max(createOrderBodyContactEmailMax),
  category: zod.enum([
    "website",
    "discord_bot",
    "repository",
    "improvement",
    "other",
  ]),
  title: zod.string().min(createOrderBodyTitleMin).max(createOrderBodyTitleMax),
  description: zod
    .string()
    .min(createOrderBodyDescriptionMin)
    .max(createOrderBodyDescriptionMax),
  paymentMethod: zod.enum(["pix", "btc", "card"]),
});

/**
 * @summary Retrieve an order by its public token
 */
export const GetOrderParams = zod.object({
  publicToken: zod.coerce.string(),
});

export const GetOrderResponse = zod.object({
  publicToken: zod.string(),
  category: zod.enum([
    "website",
    "discord_bot",
    "repository",
    "improvement",
    "other",
  ]),
  title: zod.string(),
  amountCents: zod.number(),
  paymentMethod: zod.enum(["pix", "btc", "card"]),
  status: zod.enum(["pending", "awaiting_payment", "paid", "cancelled"]),
  createdAt: zod.coerce.date(),
  pix: zod
    .union([
      zod.object({
        brCode: zod.string().describe("EMV BR Code payload string"),
        qrImageDataUrl: zod.string().describe("PNG QR code as data URL"),
        amountCents: zod.number(),
        merchantName: zod.string(),
        merchantCity: zod.string(),
      }),
      zod.null(),
    ])
    .optional(),
  btc: zod
    .union([
      zod.object({
        uri: zod.string(),
        address: zod.string(),
        amountBtc: zod.string(),
        qrImageDataUrl: zod.string(),
      }),
      zod.null(),
    ])
    .optional(),
  cardCheckoutUrl: zod.union([zod.string(), zod.null()]).optional(),
});

/**
 * Computes a server-side price estimate based on the project's category,
description length and detected complexity keywords. The estimate is
deterministic and bounded between R$ 200 and R$ 50.000. Returned as
integer cents.

 * @summary Estimate project price from category and description
 */
export const estimateQuoteBodyDescriptionMax = 2000;

export const EstimateQuoteBody = zod.object({
  category: zod.enum([
    "website",
    "discord_bot",
    "repository",
    "improvement",
    "other",
  ]),
  description: zod.string().min(1).max(estimateQuoteBodyDescriptionMax),
});

export const EstimateQuoteResponse = zod.object({
  amountCents: zod.number(),
  baseCents: zod.number(),
  factors: zod.array(
    zod.object({
      label: zod.string(),
      deltaCents: zod.number(),
    }),
  ),
});

/**
 * @summary Build a BIP21 Bitcoin URI for a given amount
 */
export const getBtcUriQueryAmountCentsMin = 500;
export const getBtcUriQueryAmountCentsMax = 5000000;

export const GetBtcUriQueryParams = zod.object({
  amountCents: zod.coerce
    .number()
    .min(getBtcUriQueryAmountCentsMin)
    .max(getBtcUriQueryAmountCentsMax),
});

export const GetBtcUriResponse = zod.object({
  uri: zod.string(),
  address: zod.string(),
  amountBtc: zod.string(),
  qrImageDataUrl: zod.string(),
});