# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderCategory } from "./orderCategory";
import type { PaymentMethod } from "./paymentMethod";

export interface CreateOrderBody {
  /**
   * @minLength 2
   * @maxLength 80
   */
  clientName: string;
  /** @maxLength 254 */
  contactEmail: string;
  category: OrderCategory;
  /**
   * @minLength 4
   * @maxLength 120
   */
  title: string;
  /**
   * @minLength 10
   * @maxLength 2000
   */
  description: string;
  paymentMethod: PaymentMethod;
}