# lib/api-zod/src/generated/types/estimateQuoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderCategory } from "./orderCategory";

export interface EstimateQuoteBody {
  category: OrderCategory;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  description: string;
}