# lib/api-zod/src/generated/types/quoteEstimate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteFactor } from "./quoteFactor";

export interface QuoteEstimate {
  amountCents: number;
  baseCents: number;
  factors: QuoteFactor[];
}